Don't error if ROLLBAR_TOKEN isn't set
